+++ lib/ui/page/home/controller.dart
@@ -49,6 +49,12 @@ class HomeController extends GetxController {
   /// Reactive [MyUser.unreadChatsCount] value.
   final Rx<int> unreadChatsCount = Rx<int>(0);
 
+  /// [Timer] for discarding any horizontal movement in a [PageView] when
+  /// non-`null`.
+  ///
+  /// Indicates currently ongoing vertical scroll of a view.
+  final Rx<Timer?> verticalScrollTimer = Rx(null);
+
   /// Authentication service to determine auth status.
   final AuthService _auth;