+++ lib/ui/page/login/controller.dart
@@ -16,6 +16,7 @@
 
 import 'package:get/get.dart';
 
+import '/api/backend/schema.dart' show CreateSessionErrorCode;
 import '/domain/model/my_user.dart';
 import '/domain/model/user.dart';
 import '/domain/service/auth.dart';
@@ -48,9 +49,6 @@ class LoginController extends GetxController {
   /// [TextFieldState] of a password text input.
   late final TextFieldState password;
 
-  /// Indicates [password] obscuring.
-  final RxBool obscurePassword = RxBool(true);
-
   /// [TextFieldState] of a recovery text input.
   late final TextFieldState recovery;
 
@@ -63,11 +61,14 @@ class LoginController extends GetxController {
   /// [TextFieldState] of a repeat password text input.
   late final TextFieldState repeatPassword;
 
-  /// Indicates [newPassword] obscuring.
+  /// Indicator whether the [password] should be obscured.
+  final RxBool obscurePassword = RxBool(true);
+
+  /// Indicator whether the [newPassword] should be obscured.
   final RxBool obscureNewPassword = RxBool(true);
 
-  /// Indicates [repeatPassword] obscuring.
-  final RxBool obscureRepeat = RxBool(true);
+  /// Indicator whether the [repeatPassword] should be obscured.
+  final RxBool obscureRepeatPassword = RxBool(true);
 
   /// [LoginViewStage] currently being displayed.
   final Rx<LoginViewStage?> stage = Rx(null);
@@ -200,7 +201,19 @@ class LoginController extends GetxController {
     } on FormatException {
       password.error.value = 'err_incorrect_password'.l10n;
     } on CreateSessionException catch (e) {
-      login.error.value = e.toMessage();
+      switch (e.code) {
+        case CreateSessionErrorCode.unknownUser:
+          login.error.value = e.toMessage();
+          break;
+
+        case CreateSessionErrorCode.wrongPassword:
+          password.error.value = e.toMessage();
+          break;
+
+        case CreateSessionErrorCode.artemisUnknown:
+          password.error.value = 'err_data_transfer'.l10n;
+          rethrow;
+      }
     } on ConnectionException {
       password.unsubmit();
       password.error.value = 'err_data_transfer'.l10n;
@@ -220,7 +233,6 @@ class LoginController extends GetxController {
     recovery.editable.value = false;
     recovery.status.value = RxStatus.loading();
     recovery.error.value = null;
-    recovery.unsubmit();
 
     _recoveryLogin = _recoveryNum = _recoveryPhone = _recoveryEmail = null;
 
@@ -283,7 +295,6 @@ class LoginController extends GetxController {
     recoveryCode.editable.value = false;
     recoveryCode.status.value = RxStatus.loading();
     recoveryCode.error.value = null;
-    recoveryCode.unsubmit();
 
     if (recoveryCode.text.isEmpty) {
       recoveryCode.editable.value = true;
@@ -328,8 +339,6 @@ class LoginController extends GetxController {
     }
 
     repeatPassword.status.value = RxStatus.empty();
-    newPassword.unsubmit();
-    repeatPassword.unsubmit();
 
     if (newPassword.text.isEmpty) {
       newPassword.error.value = 'err_input_empty'.l10n;