+++ lib/ui/page/login/view.dart
@@ -25,7 +25,7 @@ import '/ui/widget/svg/svg.dart';
 import '/ui/widget/text_field.dart';
 import 'controller.dart';
 
-/// View for logging in or recovering access on the [Routes.auth] page.
+/// View for logging in or recovering access on.
 ///
 /// Intended to be displayed with the [show] method.
 class LoginView extends StatelessWidget {
@@ -34,10 +34,10 @@ class LoginView extends StatelessWidget {
   /// Displays a [LoginView] wrapped in a [ModalPopup].
   static Future<T?> show<T>(BuildContext context) {
     return ModalPopup.show(
-      context,
-      const LoginView(),
-      contentMaxWidth: 400,
-      layoutMaxWidth: 520,
+      context: context,
+      desktopConstraints: const BoxConstraints(maxWidth: 400),
+      modalConstraints: const BoxConstraints(maxWidth: 520),
+      child: const LoginView(),
     );
   }
 
@@ -51,6 +51,7 @@ class LoginView extends StatelessWidget {
         return Obx(() {
           List<Widget> children;
 
+          // Returns a primary styled [OutlinedRoundedButton].
           Widget _primaryButton({
             Key? key,
             String? title,
@@ -68,6 +69,7 @@ class LoginView extends StatelessWidget {
             );
           }
 
+          // Returns a secondary styled [OutlinedRoundedButton].
           Widget _secondaryButton({
             Key? key,
             String? title,
@@ -85,6 +87,7 @@ class LoginView extends StatelessWidget {
             );
           }
 
+          // Returns a [Row] with [a] and [b] placed in the [Expanded] widgets.
           Row _spaced(Widget a, Widget b) {
             return Row(
               children: [
@@ -131,7 +134,7 @@ class LoginView extends StatelessWidget {
               children = [
                 Center(
                   child: Text(
-                    'label_email_confirmation_code_was_send'.l10n,
+                    'label_email_confirmation_code_was_sent'.l10n,
                     style: theme.headline3,
                     textAlign: TextAlign.center,
                   ),
@@ -148,10 +151,7 @@ class LoginView extends StatelessWidget {
                   _secondaryButton(
                     key: const Key('RecoveryCancelButton'),
                     title: 'btn_cancel'.l10n,
-                    onPressed: () {
-                      c.recovery.unsubmit();
-                      c.stage.value = null;
-                    },
+                    onPressed: () => c.stage.value = null,
                   ),
                   _primaryButton(
                     key: const Key('RecoveryNextButton'),
@@ -178,6 +178,7 @@ class LoginView extends StatelessWidget {
                   label: 'label_new_password'.l10n,
                   obscure: c.obscureNewPassword.value,
                   onSuffixPressed: c.obscureNewPassword.toggle,
+                  treatErrorAsStatus: false,
                   trailing: SvgLoader.asset(
                     'assets/icons/visible_${c.obscureNewPassword.value ? 'off' : 'on'}.svg',
                     width: 17.07,
@@ -188,10 +189,11 @@ class LoginView extends StatelessWidget {
                   key: const Key('RepeatPasswordField'),
                   state: c.repeatPassword,
                   label: 'label_repeat_password'.l10n,
-                  obscure: c.obscureRepeat.value,
-                  onSuffixPressed: c.obscureRepeat.toggle,
+                  obscure: c.obscureRepeatPassword.value,
+                  onSuffixPressed: c.obscureRepeatPassword.toggle,
+                  treatErrorAsStatus: false,
                   trailing: SvgLoader.asset(
-                    'assets/icons/visible_${c.obscureRepeat.value ? 'off' : 'on'}.svg',
+                    'assets/icons/visible_${c.obscureRepeatPassword.value ? 'off' : 'on'}.svg',
                     width: 17.07,
                   ),
                 ),
@@ -200,11 +202,7 @@ class LoginView extends StatelessWidget {
                   _secondaryButton(
                     key: const Key('RecoveryCancelButton'),
                     title: 'btn_cancel'.l10n,
-                    onPressed: () {
-                      c.recovery.unsubmit();
-                      c.recoveryCode.unsubmit();
-                      c.stage.value = null;
-                    },
+                    onPressed: () => c.stage.value = null,
                   ),
                   _primaryButton(
                     key: const Key('RecoveryNextButton'),
@@ -225,13 +223,14 @@ class LoginView extends StatelessWidget {
                     style: theme.headline3,
                   ),
                 ),
-                const SizedBox(height: 25),
+                const SizedBox(height: 32),
                 Center(
-                    child: _secondaryButton(
-                  key: const Key('RecoveryCloseButton'),
-                  title: 'btn_close'.l10n,
-                  onPressed: () => Navigator.of(context).pop(true),
-                )),
+                  child: _primaryButton(
+                    key: const Key('RecoverySuccessButton'),
+                    title: 'btn_next'.l10n,
+                    onPressed: () => c.stage.value = null,
+                  ),
+                ),
                 const SizedBox(height: 13)
               ];
               break;
@@ -249,7 +248,6 @@ class LoginView extends StatelessWidget {
                   key: const Key('UsernameField'),
                   state: c.login,
                   label: 'label_sign_in_input'.l10n,
-                  onChanged: () {},
                 ),
                 const SizedBox(height: 16),
                 ReactiveTextField(
@@ -258,6 +256,7 @@ class LoginView extends StatelessWidget {
                   label: 'label_password'.l10n,
                   obscure: c.obscurePassword.value,
                   onSuffixPressed: c.obscurePassword.toggle,
+                  treatErrorAsStatus: false,
                   trailing: SvgLoader.asset(
                     'assets/icons/visible_${c.obscurePassword.value ? 'off' : 'on'}.svg',
                     width: 17.07,
@@ -274,8 +273,11 @@ class LoginView extends StatelessWidget {
                     key: const Key('RecoveryButton'),
                     title: 'btn_forgot_password'.l10n,
                     onPressed: () {
+                      c.recovery.clear();
+                      c.recoveryCode.clear();
+                      c.newPassword.clear();
+                      c.repeatPassword.clear();
                       c.recovery.unchecked = c.login.text;
-                      c.recovery.error.value = null;
                       c.stage.value = LoginViewStage.recovery;
                     },
                   ),
@@ -284,6 +286,7 @@ class LoginView extends StatelessWidget {
               ];
               break;
           }
+
           return AnimatedSizeAndFade(
             fadeDuration: const Duration(milliseconds: 250),
             sizeDuration: const Duration(milliseconds: 250),