+++ CHANGELOG.md
@@ -17,15 +17,17 @@ All user visible changes to this project will be documented in this file. This p
     - User information auto-updating on changes ([#7], [#4]).
     - Menu:
         - Language selection ([#23]).
-    - Introduction modal ([#38]).
+    - Media panel:
+        - Reorderable buttons dock on desktop ([#9], [#6]).
+    - Introduction modal view for new users ([#38]).
 
 ### Changed
 
 - UI:
     - Media panel:
-        - Redesigned desktop interface ([#26]);
-        - Redesigned mobile interface ([#31]).
-    - Redesign login interface ([#35]).
+        - Redesigned desktop interface ([#26], [#34], [#9]);
+        - Redesigned mobile interface ([#31], [#34]).
+    - Redesigned login interface ([#35]).
 
 ### Fixed
 
@@ -34,17 +36,28 @@ All user visible changes to this project will be documented in this file. This p
 - UI:
     - Chat page:
         - Missing avatars in group creation popup ([#15], [#2]).
+    - Home page:
+        - Horizontal scroll overlapping with vertical ([#42], [#41]).
+    - Media panel:
+        - Mobile minimization gesture being too rapid ([#45], [#44]).
 
 [#2]: /../../issues/2
 [#4]: /../../issues/4
+[#6]: /../../issues/6
 [#7]: /../../pull/7
+[#9]: /../../pull/9
 [#14]: /../../pull/14
 [#15]: /../../pull/15
 [#23]: /../../pull/23
 [#26]: /../../pull/26
 [#31]: /../../pull/31
+[#34]: /../../pull/34
 [#35]: /../../pull/35
-[#35]: /../../pull/38
+[#38]: /../../pull/38
+[#41]: /../../issues/41
+[#42]: /../../pull/42
+[#44]: /../../issues/44
+[#45]: /../../pull/45